// src/product/dto/create-product.dto.ts
export class CreateProductDto {
  model: string;
  image: string;
  imageName: string;
  code: string;
  price: string;
  yearId: number;
  brandId: number;
  typeId: number;
  cpuId: number;
  colorId: number;
  graphicsId: number;
  osId: number;
  resolutionId: number;
  ramTypeId: number;
  ram: string;
  core: number;
  diagonal: string;
  sizeHD: string;
  refreshRate: string;
  weight: string;
  thickness: string;
  cpuModel: string;
  graphicsModel: string;
  url: string;
}